chore: add csharp sdk tests; fix some issues with csharp client

+++ flipt-csharp/src/DTOs/BatchEvaluationResponse.cs
@@ -15,7 +15,7 @@ public class BatchEvaluationResponse
     public required Response[] Responses { get; set; }
 
     [Required]
-    public int RequestDurationMillis { get; set; }
+    public float RequestDurationMillis { get; set; }
 }
