+++ flipt-csharp/src/DTOs/BooleanEvaluationResponse.cs
@@ -17,7 +17,7 @@ public class BooleanEvaluationResponse
     [Required]
     public DateTime Timestamp { get; set; }
     [Required]
-    public int RequestDurationMillis { get; set; }
+    public float RequestDurationMillis { get; set; }
     [Required]
     public Reason Reason { get; set; }
 }
